Don't set store on factory, pass it around instead
This is to prepare for allowing forking the store & operating in an 
isolated context.

Plan is as follows:

    var person = this.store.find("person", 1);
    var isolated = this.store.fork();
    var otherPerson = isolated.store.find("person", 1);


`otherPerson` is now isolated from `person` and all associations / finds
that it kicks off are also isolated & guaranteed to not be from the main
store's cache etc...

Means you can operate on a copy of the tree & then save at the end or 
discard changes if you don't want to apply them.