WIP - Return PromiseProxy from finds
This has the potential to be confusing as the proxy won't support #save etc... you'll
have to know to drop through to the proxies content to save etc...

More confusion because the proxy isn't put in the cache, the underlying record is.
So a 2nd call to find will return an actual record and not a proxy!

Can we unwrap this when it's resolved? Not if it's been assigned to something.
Indexed collection could replace the record in its cache when it's resolved?

If we had method-missing then all would be fine, as the issue comes when calling
functions on the proxy/record. We could implement methods on the proxy for the 
common functions (save, buildFirebaseReference etc...) and forward them to its content
but that's going to be very confusing if you add your own methods!