Wait for collection to be loaded before resolving Store#fetch
This shouldn’t break anything, but I’ll bump to v0.2.0 just in-case when 
it’s released.

Main use case is for indexed collections:

```javascript
this.store.fetch(“person”, {
  collection: “index”, 
  path: “/foo/bar”
});
```

Previously that would resolve as soon as the index was loaded
meaning you couldn’t rely on its content being populated.

Now that will only resolve when the items are loaded too.